+++ crates/ai00-server/src/api/oai/chat.rs
@@ -120,7 +120,7 @@ struct ChatRequest {
     stop: Array<String>,
     stream: bool,
     #[serde(alias = "logit_bias")]
-    bias: HashMap<u16, f32>,
+    bias: HashMap<u32, f32>,
     bnf_schema: Option<String>,
     #[serde(alias = "sampler_override")]
     sampler: Option<SamplerParams>,
